Move no missing UnityEvent targets to scenevalidationbasefixture class
